# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Vending machine management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /machines:
    get:
      operationId: listMachines
      tags: [machines]
      summary: List all machine definitions
      responses:
        "200":
          description: List of machines
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Machine"
    post:
      operationId: createMachine
      tags: [machines]
      summary: Create a machine definition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MachineInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Machine"
  /machines/{id}:
    get:
      operationId: getMachine
      tags: [machines]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Machine
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Machine"
    patch:
      operationId: updateMachine
      tags: [machines]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MachineUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Machine"
    delete:
      operationId: deleteMachine
      tags: [machines]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Products ──────────────────────────────────────────────────────────────
...
  /machine-loads:
    get:
      operationId: listMachineLoads
      tags: [machine-loads]
      summary: List machine load records
      parameters:
        - name: clientId
          in: query
          schema:
            type: ["integer", "null"]
        - name: clientMachineId
          in: query
          schema:
            type: ["integer", "null"]
        - name: operatorId
          in: query
          schema:
            type: ["integer", "null"]
        - name: dateFrom
          in: query
          schema:
            type: ["string", "null"]
        - name: dateTo
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Machine load records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MachineLoad"
    post:
      operationId: createMachineLoad
      tags: [machine-loads]
      summary: Record a machine load (operator refills machine)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MachineLoadInput"
      responses:
        "201":
          description: Load recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MachineLoad"

  # ── Schedules ─────────────────────────────────────────────────────────────
...
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      summary: List expenses
      parameters:
        - name: clientId
          in: query
          schema:
            type: ["integer", "null"]
        - name: dateFrom
          in: query
          schema:
            type: ["string", "null"]
        - name: dateTo
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Create an expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /expenses/{id}:
    patch:
      operationId: updateExpense
      tags: [expenses]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
    delete:
      operationId: deleteExpense
      tags: [expenses]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Reports ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Machine:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [coffee, vending]
        brand:
          type: ["string", "null"]
        model:
          type: ["string", "null"]
        containerCount:
          type: ["integer", "null"]
        rowCount:
          type: ["integer", "null"]
        chuteCount:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - type
        - createdAt
    MachineInput:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [coffee, vending]
        brand:
          type: ["string", "null"]
        model:
          type: ["string", "null"]
        containerCount:
          type: ["integer", "null"]
        rowCount:
          type: ["integer", "null"]
        chuteCount:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
      required:
        - name
        - type
    MachineUpdate:
      type: object
      properties:
        name:
          type: ["string", "null"]
        brand:
          type: ["string", "null"]
        model:
          type: ["string", "null"]
        containerCount:
          type: ["integer", "null"]
        rowCount:
          type: ["integer", "null"]
        chuteCount:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]

    # ── Product ──────────────────────────────────────────────────────────────
...
    MachineLoad:
      type: object
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: string
        clientMachineId:
          type: integer
        machineNumber:
          type: string
        machineName:
          type: string
        operatorId:
          type: ["integer", "null"]
        operatorName:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/MachineLoadItem"
        totalRevenue:
          type: number
        totalCost:
          type: number
        totalProfit:
          type: number
        createdAt:
          type: string
      required:
        - id
        - clientId
        - clientName
        - clientMachineId
        - machineNumber
        - machineName
        - items
        - totalRevenue
        - totalCost
        - totalProfit
        - createdAt
    MachineLoadItem:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: number
        purchasePrice:
          type: number
        salePrice:
          type: number
      required:
        - productId
        - productName
        - quantity
        - purchasePrice
        - salePrice
    MachineLoadInput:
      type: object
      properties:
        clientMachineId:
          type: integer
        operatorId:
          type: ["integer", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/MachineLoadItemInput"
      required:
        - clientMachineId
        - items
    MachineLoadItemInput:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: number
      required:
        - productId
        - quantity

    # ── Schedule ─────────────────────────────────────────────────────────────
...
    Expense:
      type: object
      properties:
        id:
          type: integer
        clientId:
          type: ["integer", "null"]
        clientName:
          type: ["string", "null"]
        category:
          type: string
          enum: [electricity, rent, repairs, salary, other]
        amount:
          type: number
        description:
          type: ["string", "null"]
        isRecurring:
          type: boolean
        date:
          type: string
        createdAt:
          type: string
      required:
        - id
        - category
        - amount
        - isRecurring
        - date
        - createdAt
    ExpenseInput:
      type: object
      properties:
        clientId:
          type: ["integer", "null"]
        category:
          type: string
          enum: [electricity, rent, repairs, salary, other]
        amount:
          type: number
        description:
          type: ["string", "null"]
        isRecurring:
          type: boolean
        date:
          type: string
      required:
        - category
        - amount
        - isRecurring
        - date
    ExpenseUpdate:
      type: object
      properties:
        category:
          type: ["string", "null"]
        amount:
          type: ["number", "null"]
        description:
          type: ["string", "null"]
        isRecurring:
          type: ["boolean", "null"]
        date:
          type: ["string", "null"]

    # ── Reports ──────────────────────────────────────────────────────────────
...