# lib/api-zod/src/generated/types/expenseUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vending machine management API
 * OpenAPI spec version: 0.1.0
 */

export interface ExpenseUpdate {
  /** @nullable */
  category?: string | null;
  /** @nullable */
  amount?: number | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  isRecurring?: boolean | null;
  /** @nullable */
  date?: string | null;
}